switch to experimental higher half kernel model and some changes
